+++ HISTORY.md
@@ -1,6 +1,7 @@
 # Release History
 
 ## css-sanitizer x.x.x (not yet released)
+* Harden security by disallowsing CSS custom properties for `border-color`
 
 ## css-sanitizer 5.2.0 (2024-04-03)
 * Add support for CSS custom properties in color attributes